+++ src/Commands/Enums/AzureACSPrincipalScope.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PnP.PowerShell.Commands.Enums
+{
+    public enum AzureACSPrincipalScope
+    {
+        List,
+        Web,
+        Site,
+        Tenant,
+        All
+    }
+}